this commit will adjust the wording in the tour guide to clarify to users what each aspect does
